fixed main class name in build script.
added support for comments in scxml file to include geometric information.
the layout algorithm currently don't support nodes with fixed position so the geometric
information is currently ignored.

fixed problem with adding corners in edges those coordinates are relative to some
other object.

fixed problem with integration of add/remove of corners and undo/redo functionality.

layout in popup menu is now applied only to the selected node.

changed default color of edges to gray so that the labels are more visible.

changed the use of the mouse wheel to scroll the graph. now working both on
vertical and horizontal scroll (using the shift as modifier).

git-svn-id: http://scxmlgui.googlecode.com/svn/trunk@4 cf63a8c4-81e5-8ad9-1f9c-79d1096461f8