added zoom control with CTRL-PAGE-UP/DOWN.
now the root scxml node cannot be deleted.

connected the change tracker that marks the modified flag of the current graph to the
undo manager (if the undo-manager can undo then the graph is modified).

added support to make individual cells un-deletable in addition to un-movable.

extended scxml listener to allow for update of the list of events while the user is navigating the history of events.

added export to dot format (graphviz).

git-svn-id: http://scxmlgui.googlecode.com/svn/trunk@8 cf63a8c4-81e5-8ad9-1f9c-79d1096461f8