fixed modified flag tracking when the number of modifications overflows the capacity set in the undo manager.
fixed bug in detecting when a node is un-movable.

fixed problem when reading an scxml file that caused all spaces in a label to be removed.

added tracking of index of unmodified state in undo manager to fix a problem with the
modified flag when the number of undo steps overflows the max limit.

modified method that returns the bounding box of the content of a node to include also
the space used by labels.

git-svn-id: http://scxmlgui.googlecode.com/svn/trunk@11 cf63a8c4-81e5-8ad9-1f9c-79d1096461f8