removed the setAlwaysOnTop property to true for the various property editors because the right behavior is obtained by setting the editor as children of the main window.
improved display of outsourced content: error message if a node is not found, better file chooser to lookup a file not found)
fixed: forgot to pass the screen coord to the edge editor.
changed history of recent files to 9 otherwise not enough mnemonics.

fixed problem with layout of folded clusters.



git-svn-id: http://scxmlgui.googlecode.com/svn/trunk@37 cf63a8c4-81e5-8ad9-1f9c-79d1096461f8